daemon: de-duplicate adding devices
Two almost identical code paths, one for devices already present at startup,
one for devices hotplugged at runtime. The code paths differed in some ways
but it's hard to tell from the history whether that was intentional or not.

De-duplicate that code path and use the same method for both paths to make
sure our behavior is consistent.